Add a bunch of neat stuff for the IR sensor
Cleaned up some of the configuration work, and added a setting (default)
for running the LED on a 50ms on, 450ms off cycle, so that the poor
person testing this thing doesn't have to keep pressing the button.
Thought about it and realized Patrizzle was 100% right about that being
the right thing to do.

Hopefully the timer code isn't too much of a dog's breakfast.