Cleaned up IR-LED transmit code a ton
Used app_timer instead of whatever black evil magic i was using before.
Things are much better now!!

Removed that weird bit of code where we were just waiting for the pwm
setting to change when using continuous transmission mode

added _main to the end of the main loop update function

NOTE: Using the logic analyser I noticed that the frequency we're
pulsing the LED is actually about 38461kHz. It's possible to get the
timing right by screwing around with app_pwm, and I plan to do so in
another commit, on a different branch.