Added some magic to create timers based on ns
The PWM timing for the IR LED was only able to give us about 38.5 kHz.
This is technically sufficient, but....
If you add (but don't take away!) a little bit of extra code to some of
the drivers, we can create timers that are based on ns instead of us,
and are therefore more precise. Using these changes I was able to get
the IR pulsing at around 38.15kHz, Tested with a logic analyzer.

This isn't perfect, but it's better!