Add application to read data and broadcast it
Don't merge this branch yet!! In order to get the ANT transfer working
properly, we had to hack mm_ant_control.c to notify grideye.c directly
when a message is transferred.

Wait for the branch with registering ANT listeners to be merged first,
then merge master back into this branch, and clean things up.

This commit has all the code needed to read data off the grid-eye and
broadcast the data over ANT at 1 fps in a compressed format. The code to
decompress the data will be made available in the near future.