Add Power duty cycling, median filtering
Added a whole rigmarole for acquiring samples at a certain rate, and
turning the lidar off otherwise. This seems to work pretty well,
although at the few distances I tested it seems like there was up to a
20cm offset (measuring 147cm when it should have measured 167). However,
given that we're looking for relative changes in distance measured, this
should work pretty well anyway.

Speaking of that, this adds some functionality for median filtering
samples, and setting a flag when we detect a significant change in the
filtered distance. The most recent filtered distance and the flag are
accessible by public functions.