Initial commit for RGB LED control
With these changes you can:
 - Set the RGB colour of the strip with a hexcode
 - Set the individual R G and B values of the strip
 - Set the amount of time the strip is on and off for, in a cycle
   (e.g., on for 500ms, then off for 500ms, repeat.)

The pins being used for this include some of the pins designated for the
third PIR sensor, but we're running out of GPIOs and I'm assuming that
we probably won't ever need a third PIR sensor on a roadside node, which
have the LED strips.

Another feature still needed for this module is a bit of code and
hardware to make sure that the battery pack always stays on. Otherwise
it turns off after like 10-30s of low current draw.