Add test, power bank control, bug fixes
Added a test wherein the LEDs are set to white for 5 seconds on startup.
This is kinda nice since it's a good test of wiring, brightness, and
timing.
Also added some code for tricking the power bank into staying on all the
time. Much more information is available on google drive, but in a
nutshell it explains the default timing values, and why this may not
have been necessary! (The power bank may have tricked ME!)

Also fixed some bugs so this code is much more stable now.