Add hardware test file
Added a helpful file for running tests of different hardware
configurations! The next update I'd want to make to this is to
incorporate reading the dip switches to check what the hardware should
be rather than passing it as an argument.

Information about the tests for the different hardware configurations
are on the drive under Capstone > Hardware General > Full Node Hardware
Testing.

Also includes a few other small changes:
 - Added the LIDAR header location to the right spot in include.mk
 - Changed the LIDAR and PIR code to use a dynamic variable for LED
   debugging, so it can be set on or off on init.
 - Cleaned up a few other nitpicky things in the LIDAR and PIR code