Cleanup, change PIR to use scheduler
Cleaned up a bunch of unused files left over from sensor testing,
including the PWM and PPI apps (we use low-power PWM for the LED strip),
as well as the alternative PIR code and IR LED transmit/receive code.

Also changed the pir_st_00081 code to use the scheduler instead of
having a main callback. Verified that it works as well as before using
debugging and letting the system run for a while.